+++ alignak_webui/objects/item_actions.py
@@ -1,0 +1,76 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# Many functions need to use protected members of a base class
+# pylint: disable=protected-access
+# Attributes need to be defined in constructor before initialization
+# pylint: disable=attribute-defined-outside-init
+
+# Copyright (c) 2015-2016:
+#   Frederic Mohier, frederic.mohier@gmail.com
+#
+# This file is part of (WebUI).
+#
+# (WebUI) is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# (WebUI) is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with (WebUI).  If not, see <http://www.gnu.org/licenses/>.
+
+"""
+    This module contains the classes used to manage the backend actions objects.
+"""
+
+from logging import getLogger, INFO
+
+# Import the backend interface class
+from alignak_webui.objects.element import Element
+
+# Set logger level to INFO, this to allow global application DEBUG logs without being spammed... ;)
+logger = getLogger(__name__)
+logger.setLevel(INFO)
+
+
+class ActionAcknowledge(Element):
+    """
+    Object representing an actionacknowledge
+    """
+    _count = 0
+    # Next value used for auto generated id
+    _next_id = 1
+    # _type stands for Backend Object Type
+    _type = 'actionacknowledge'
+    # _cache is a list of created objects
+    _cache = {}
+
+
+class ActionDowntime(Element):
+    """
+    Object representing an actiondowntime
+    """
+    _count = 0
+    # Next value used for auto generated id
+    _next_id = 1
+    # _type stands for Backend Object Type
+    _type = 'actiondowntime'
+    # _cache is a list of created objects
+    _cache = {}
+
+
+class ActionForceCheck(Element):
+    """
+    Object representing an actionforcecheck
+    """
+    _count = 0
+    # Next value used for auto generated id
+    _next_id = 1
+    # _type stands for Backend Object Type
+    _type = 'actionforcecheck'
+    # _cache is a list of created objects
+    _cache = {}
