+++ alignak_webui/objects/item_service.py
@@ -1,0 +1,159 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# Many functions need to use protected members of a base class
+# pylint: disable=protected-access
+# Attributes need to be defined in constructor before initialization
+# pylint: disable=attribute-defined-outside-init
+
+# Copyright (c) 2015-2016:
+#   Frederic Mohier, frederic.mohier@gmail.com
+#
+# This file is part of (WebUI).
+#
+# (WebUI) is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# (WebUI) is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with (WebUI).  If not, see <http://www.gnu.org/licenses/>.
+
+"""
+    This module contains the classes used to manage the application objects with the data manager.
+"""
+
+from logging import getLogger, INFO
+
+# Import the backend interface class
+
+# Set logger level to INFO, this to allow global application DEBUG logs without being spammed... ;)
+from alignak_webui.objects.element import Element
+
+logger = getLogger(__name__)
+logger.setLevel(INFO)
+
+
+class Service(Element):
+    """
+    Object representing a service
+    """
+    _count = 0
+    # Next value used for auto generated id
+    _next_id = 1
+    # _type stands for Backend Object Type
+    _type = 'service'
+    # _cache is a list of created objects
+    _cache = {}
+
+    def __new__(cls, params=None, date_format='%a, %d %b %Y %H:%M:%S %Z'):
+        """
+        Create a new user
+        """
+        return super(Service, cls).__new__(cls, params, date_format)
+
+    def _create(self, params, date_format):
+        """
+        Create a service (called only once when an object is newly created)
+        """
+        self._linked_host = 'host'
+        self._linked_check_command = 'command'
+        self._linked_event_handler = 'command'
+        self._linked_check_period = 'timeperiod'
+        self._linked_notification_period = 'timeperiod'
+        self._linked_servicegroups = 'servicegroup'
+        self._linked_users = 'user'
+        self._linked_usergroups = 'usergroup'
+
+        super(Service, self)._create(params, date_format)
+
+    def _update(self, params=None, date_format='%a, %d %b %Y %H:%M:%S %Z'):
+        """
+        Update a service (called every time an object is updated)
+        """
+        super(Service, self)._update(params, date_format)
+
+    def __init__(self, params=None):
+        """
+        Initialize a service (called every time an object is invoked)
+        """
+        super(Service, self).__init__(params)
+
+    @property
+    def check_command(self):
+        """ Return linked object """
+        return self._linked_check_command
+
+    @property
+    def event_handler(self):
+        """ Return linked object """
+        return self._linked_event_handler
+
+    @property
+    def host(self):
+        """ Return linked object """
+        return self._linked_host
+
+    @property
+    def check_period(self):
+        """ Return linked object """
+        return self._linked_check_period
+
+    @property
+    def notification_period(self):
+        """ Return linked object """
+        return self._linked_notification_period
+
+
+class ServiceGroup(Element):
+    """
+    Object representing a servicegroup
+    """
+    _count = 0
+    # Next value used for auto generated id
+    _next_id = 1
+    # _type stands for Backend Object Type
+    _type = 'servicegroup'
+    # _cache is a list of created objects
+    _cache = {}
+
+    def __new__(cls, params=None, date_format='%a, %d %b %Y %H:%M:%S %Z'):
+        """
+        Create a new servicegroup
+        """
+        return super(ServiceGroup, cls).__new__(cls, params, date_format)
+
+    def _create(self, params, date_format):
+        """
+        Create a servicegroup (called only once when an object is newly created)
+        """
+        self._linked_servicegroup_members = 'servicegroup'
+        self._linked_members = 'service'
+
+        super(ServiceGroup, self)._create(params, date_format)
+
+    def _update(self, params=None, date_format='%a, %d %b %Y %H:%M:%S %Z'):
+        """
+        Update a servicegroup (called every time an object is updated)
+        """
+        super(ServiceGroup, self)._update(params, date_format)
+
+    def __init__(self, params=None):
+        """
+        Initialize a servicegroup (called every time an object is invoked)
+        """
+        super(ServiceGroup, self).__init__(params)
+
+    @property
+    def members(self):
+        """ Return linked object """
+        return self._linked_members
+
+    @property
+    def servicegroup_members(self):
+        """ Return linked object """
+        return self._linked_servicegroup_members
