+++ alignak_webui/plugins/livestate/livestate.py
@@ -23,15 +23,12 @@
     Plugin Livestate
 """
 
-import time
-import json
-
 from collections import OrderedDict
-
 from logging import getLogger
-from bottle import request, response, redirect, template
 
-from alignak_webui.objects.item import Item
+from bottle import request, redirect
+
+from alignak_webui import _
 from alignak_webui.plugins.common.common import get_widget, get_table, get_table_data
 
 logger = getLogger(__name__)
