+++ test/test_backend_alignak.py
@@ -20,27 +20,28 @@
 # along with (WebUI).  If not, see <http://www.gnu.org/licenses/>.
 # import the unit testing module
 
+from __future__ import print_function
+
 import os
-import json
-import time
 import shlex
-import unittest2
 import subprocess
+import time
 
-from nose import with_setup
-from nose.tools import *
-
+import unittest2
 from alignak_backend_client.client import BACKEND_PAGINATION_LIMIT, BACKEND_PAGINATION_DEFAULT
+
 from alignak_webui.objects.backend import BackendConnection
 
 pid = None
 backend_address = "http://127.0.0.1:5000/"
+
+
 # backend_address = "http://94.76.229.155:80"
 
 
-def setup_module(module):
-    print ("")
-    print ("start alignak backend")
+def setup_module():
+    print("")
+    print("start alignak backend")
 
     global pid
     global backend_address
@@ -52,7 +53,8 @@ def setup_module(module):
 
         # Delete used mongo DBs
         exit_code = subprocess.call(
-            shlex.split('mongo %s --eval "db.dropDatabase()"' % os.environ['ALIGNAK_BACKEND_MONGO_DBNAME'])
+            shlex.split(
+                'mongo %s --eval "db.dropDatabase()"' % os.environ['ALIGNAK_BACKEND_MONGO_DBNAME'])
         )
         assert exit_code == 0
         time.sleep(1)
@@ -61,61 +63,61 @@ def setup_module(module):
         pid = subprocess.Popen(
             shlex.split('alignak_backend')
         )
-        print ("PID: %s" % pid)
+        print("PID: %s" % pid)
         time.sleep(3)
 
-        print ("")
-        print ("populate backend content")
+        print("")
+        print("populate backend content")
         exit_code = subprocess.call(
             shlex.split('alignak_backend_import --delete cfg/default/_main.cfg')
         )
         assert exit_code == 0
 
-def teardown_module(module):
-    print ("")
-    print ("stop applications backend")
+
+def teardown_module():
+    print("")
+    print("stop applications backend")
 
     if backend_address == "http://127.0.0.1:5000/":
         global pid
         pid.kill()
 
 
-class test_1_creation(unittest2.TestCase):
-
+class TestCreation(unittest2.TestCase):
     def setUp(self):
-        print ""
+        print("")
 
     def tearDown(self):
-        print ""
+        print("")
 
     def test_creation(self):
-        print "--- creation"
+        print("--- creation")
 
         be = BackendConnection(backend_address)
         assert be
-        print "Backend object:", be
+        print("Backend object:", be)
 
         be2 = BackendConnection(backend_address)
         assert be2
-        print "Backend object:", be2
+        print("Backend object:", be2)
 
         # Objet is a singleton
         assert be == be2
 
-class test_2_login(unittest2.TestCase):
 
+class TestLogin(unittest2.TestCase):
     def setUp(self):
-        print ""
+        print("")
 
     def tearDown(self):
-        print ""
+        print("")
 
     def test_login(self):
-        print "--- login"
+        print("--- login")
 
         be = BackendConnection(backend_address)
         assert be
-        print "Backend object:", be
+        print("Backend object:", be)
 
         be.login('admin', 'fake')
         assert not be.connected
@@ -124,102 +126,101 @@ class test_2_login(unittest2.TestCase):
         assert be.connected
 
 
-class test_3_get(unittest2.TestCase):
-
+class TestGet(unittest2.TestCase):
     def setUp(self):
-        print "Login..."
+        print("Login...")
         self.be = BackendConnection(backend_address)
         assert self.be
         self.be.login('admin', 'admin')
         assert self.be.connected
 
-        print "Logged in"
+        print("Logged in")
 
     def tearDown(self):
-        print ""
+        print("")
 
     def test_count(self):
-        print "--- count"
+        print("--- count")
 
         # Count all users
         result = self.be.count('user')
-        print "Result: %s", result
+        print("Result: %s", result)
         self.assertEqual(result, 4)
 
-        parameters = {'where': {"name":"admin"}}
+        parameters = {'where': {"name": "admin"}}
         result = self.be.count('user', parameters)
-        print "Result: %s", result
+        print("Result: %s", result)
         self.assertEqual(result, 1)
 
-        parameters = {'where': {"name":"fake"}}
+        parameters = {'where': {"name": "fake"}}
         result = self.be.count('user', parameters)
-        print "Result: %s", result
-        self.assertEqual(result, 0) # Not found !
+        print("Result: %s", result)
+        self.assertEqual(result, 0)  # Not found !
 
         # Get admin user
-        parameters = {'where': {"name":"admin"}}
+        parameters = {'where': {"name": "admin"}}
         result = self.be.get('user', parameters)
-        print result
-        self.assertEqual(len(result), 1)    # Only 1 is admin
+        print(result)
+        self.assertEqual(len(result), 1)  # Only 1 is admin
 
         result = self.be.count('user', result[0]['_id'])
-        print "Result: %s", result
+        print("Result: %s", result)
         self.assertEqual(result, 1)
 
     def test_get(self):
-        print "--- get"
+        print("--- get")
 
         # Get all users
         result = self.be.get('user')
-        print "%s users: " % len(result)
+        print("%s users: " % len(result))
         for user in result:
             self.assertIn('name', user)
-            self.assertIn('_total', user)    # Each element has an extra _total attribute !
-            print " - %s (one out of %d)" % (user['name'], user['_total'])
-        self.assertEqual(len(result), 4)        # Default configuration has 4 users
+            self.assertIn('_total', user)  # Each element has an extra _total attribute !
+            print(" - %s (one out of %d)" % (user['name'], user['_total']))
+        self.assertEqual(len(result), 4)  # Default configuration has 4 users
 
-        parameters = {'where': {"name":"fake"}}
+        parameters = {'where': {"name": "fake"}}
         result = self.be.get('user', parameters)
-        print result
-        self.assertEqual(len(result), 0)    # Not found
+        print(result)
+        self.assertEqual(len(result), 0)  # Not found
 
-        parameters = {'where': {"name":"admin"}}
+        parameters = {'where': {"name": "admin"}}
         result = self.be.get('user', parameters)
-        print result
-        self.assertEqual(len(result), 1)    # Only 1 is admin
+        print(result)
+        self.assertEqual(len(result), 1)  # Only 1 is admin
         admin_id = result[0]['_id']
-        print "Administrator id:", admin_id
+        print("Administrator id:", admin_id)
 
         result = self.be.get('user', result[0]['_id'])
-        print "Result: %s", result
-        self.assertEqual(len(result), 1)    # Only 1 is admin
+        print("Result: %s", result)
+        self.assertEqual(len(result), 1)  # Only 1 is admin
         self.assertEqual(result[0]['_id'], admin_id)
 
         # Directly address object in the backend
         result = self.be.get('user/' + result[0]['_id'])
-        print "--- Result: %s", result
-        self.assertEqual(len(result), 39)    # 39 attributes in the result
+        print("--- Result: %s", result)
+        self.assertEqual(len(result), 39)  # 39 attributes in the result
         self.assertEqual(result['_id'], admin_id)
 
     def test_get_all(self):
-        print "--- get all"
+        print("--- get all")
 
         # Get one page of services
         result = self.be.get('service')
-        print "Backend pagination default:", BACKEND_PAGINATION_DEFAULT
-        print "Backend pagination limit:", BACKEND_PAGINATION_LIMIT
+        print("Backend pagination default:", BACKEND_PAGINATION_DEFAULT)
+        print("Backend pagination limit:", BACKEND_PAGINATION_LIMIT)
 
-        print "%s services: " % len(result)
+        print("%s services: " % len(result))
         for service in result:
-            print " - %s" % service['name']
+            print(" - %s" % service['name'])
         assert len(result) == BACKEND_PAGINATION_LIMIT
-        # Should be DEFAULT and not LIMIT ... See https://github.com/Alignak-monitoring-contrib/alignak-backend/issues/52
+        # Should be DEFAULT and not LIMIT ...
+        # See https://github.com/Alignak-monitoring-contrib/alignak-backend/issues/52
         # assert len(result) == BACKEND_PAGINATION_DEFAULT # Default backend pagination
 
         # Get all services
         result = self.be.get('service', all_elements=True)
-        print "%s services: " % len(result)
+        print("%s services: " % len(result))
         for service in result:
-            print " - %s" % service['name']
+            print(" - %s" % service['name'])
         self.assertEqual(len(result), 94)
-
